Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,9 +1,8 @@
 #include <stdio.h>
 #include <math.h>
-
 int main() {
     int a, b, sum, multiply, subtraction;
-    int division;
+    float division;
     char d;
     
     // Fix: Add spaces to allow proper reading of integers and operator
